Split ability and projectile io to avoid circular imports.
