+++ src/items/bullet_weapons.py
@@ -7,7 +7,7 @@ import images
 import settings
 import sounds
 from abilities import ProjectileAbilityData
-from data.input_output import load_projectile_data
+from data.projectiles_io import load_projectile_data
 from model import DynamicObject
 from mods import ModLocation, ItemObject, ModData, Mod
 from tilemap import ObjectType
