+++ src/items/laser_weapons.py
@@ -4,7 +4,7 @@ from pygame.surface import Surface
 import images
 import sounds
 from abilities import ProjectileAbilityData
-from data.input_output import load_projectile_data
+from data.projectiles_io import load_projectile_data
 from mods import ModLocation, ItemObject, ModData, Mod
 from tilemap import ObjectType
